# frontend/src/lib/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ExpertModelUpdateRequest": {
        "properties": {
          "assembly_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assembly Enabled"
          },
          "data_generation_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Generation Enabled"
          },
          "demo_prompt": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demo Prompt"
          },
          "demo_prompt_label": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demo Prompt Label"
          },
          "domain": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "simulator": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Simulator"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "active",
                  "disabled",
                  "invalid"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "ExpertModelUpdateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PierNet__new_synth__schemas__ExpertGenerateRequest": {
        "properties": {
          "input_dim": {
            "anyOf": [
              {
                "maximum": 1000000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Dim"
          },
          "model_id": {
            "maxLength": 180,
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "prompt": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "scenario": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Scenario",
            "type": "string"
          }
        },
        "required": [
          "model_id",
          "scenario",
          "prompt"
        ],
        "title": "ExpertGenerateRequest",
        "type": "object"
      },
...
      "ProjectSnapshot": {
        "properties": {
          "artifacts": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifacts"
          },
          "can_chat": {
            "default": false,
            "title": "Can Chat",
            "type": "boolean"
          },
          "can_run": {
            "default": false,
            "title": "Can Run",
            "type": "boolean"
          },
          "compatibility": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compatibility"
          },
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "current_stage": {
            "title": "Current Stage",
            "type": "string"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "expert": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expert"
          },
          "goal": {
            "title": "Goal",
            "type": "string"
          },
          "inspection": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspection"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "recommended_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Prompt"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageSnapshot"
            },
            "title": "Stages",
            "type": "array"
          },
          "status": {
            "enum": [
              "draft",
              "ready",
              "running",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "number"
          }
        },
        "required": [
          "project_id",
          "name",
          "goal",
          "status",
          "current_stage",
          "created_at",
          "updated_at",
          "stages"
        ],
        "title": "ProjectSnapshot",
        "type": "object"
      },
...
      "StageSnapshot": {
        "properties": {
          "finished_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "progress": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "waiting",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "message"
        ],
        "title": "StageSnapshot",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkflowSnapshot": {
        "properties": {
          "artifacts": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifacts"
          },
          "can_define": {
            "default": false,
            "title": "Can Define",
            "type": "boolean"
          },
          "can_generate": {
            "default": false,
            "title": "Can Generate",
            "type": "boolean"
          },
          "can_open_training": {
            "default": false,
            "title": "Can Open Training",
            "type": "boolean"
          },
          "cancel_requested": {
            "default": false,
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "current_step": {
            "enum": [
              "source",
              "definition",
              "generation",
              "complete"
            ],
            "title": "Current Step",
            "type": "string"
          },
          "definition": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "status": {
            "enum": [
              "draft",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "number"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id",
          "name",
          "status",
          "current_step",
          "created_at",
          "updated_at"
        ],
        "title": "WorkflowSnapshot",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Piern Unified API",
    "version": "3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/expert-models/upload": {
      "post": {
        "operationId": "upload_expert_model_api_expert_models_upload_post",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "maxLength": 180,
              "minLength": 1,
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Upload Expert Model Api Expert Models Upload Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Expert Model",
        "tags": [
          "expert-models"
        ]
      }
    },
    "/api/expert-models/{model_id}": {
      "delete": {
        "operationId": "delete_expert_model_api_expert_models__model_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Expert Model",
        "tags": [
          "expert-models"
        ]
      },
      "get": {
        "operationId": "get_expert_model_api_expert_models__model_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Expert Model Api Expert Models  Model Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Expert Model",
        "tags": [
          "expert-models"
        ]
      },
      "patch": {
        "operationId": "update_expert_model_api_expert_models__model_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpertModelUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Expert Model Api Expert Models  Model Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Expert Model",
        "tags": [
          "expert-models"
        ]
      }
    },
    "/api/expert-models/{model_id}/validate": {
      "post": {
        "operationId": "validate_expert_model_api_expert_models__model_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Validate Expert Model Api Expert Models  Model Id  Validate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Expert Model",
        "tags": [
          "expert-models"
        ]
      }
    },
...
    "/api/new-synth/experts/upload": {
      "post": {
        "operationId": "upload_expert_api_new_synth_experts_upload_post",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "maxLength": 180,
              "minLength": 1,
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-File-Name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-File-Name"
            }
          },
          {
            "in": "cookie",
            "name": "piern_new_synth_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Piern New Synth Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Expert",
        "tags": [
          "new-synth"
        ]
      }
    },
...
    "/api/new-synth/workflows/{workflow_id}/source/expert": {
      "post": {
        "operationId": "generate_expert_source_api_new_synth_workflows__workflow_id__source_expert_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "piern_new_synth_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Piern New Synth Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PierNet__new_synth__schemas__ExpertGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Expert Source",
        "tags": [
          "new-synth"
        ]
      }
    },
...
    "/api/simulation/expert-models/upload": {
      "post": {
        "description": "上传专家模型 .py 文件或带 manifest 的压缩包；入口必须满足 predict(inputs: list[float]).",
        "operationId": "upload_expert_model_api_simulation_expert_models_upload_post",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "maxLength": 180,
              "minLength": 1,
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Expert Model"
      }
    },
...
    "/api/studio/projects/{project_id}/expert": {
      "post": {
        "operationId": "upload_expert_api_studio_projects__project_id__expert_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-File-Name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-File-Name"
            }
          },
          {
            "in": "cookie",
            "name": "piern_studio_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Piern Studio Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Expert",
        "tags": [
          "studio"
        ]
      }
    },
...
  }
}